+++ public/index.js
@@ -1,6 +1,5 @@
 'use strict'
 
-exports.generate_key_pair = require('./generate_eos_signature.js')
-exports.get_public_key = require('./public_key_from_private.js')
-exports.sign_hash = require('./new_eos_keys.js')
-exports.verify_signature = require('./verify_eos_signature.js')
+exports.sign_txn = require('./sign_txn.js')
+exports.public_key_from_private = require('./public_key_from_private.js')
+exports.new_eos_keys = require('./new_eos_keys.js')