+++ public/new_eos_keys.js
@@ -1,8 +1,8 @@
 'use strict'
-
-const generate_key_pair = require('isomorphic-secp256k1/public/generate_key_pair')
+const { get_public_key } = require('universal-ecdsa')
 const private_key_to_wif = require('../private/private_key_to_wif')
 const public_key_to_wif = require('../private/public_key_to_wif')
+const random_bytes = require('../private/random_bytes')
 
 /**
  * An EOS wallet import formatted (WIF) public & private key pair.
@@ -16,6 +16,7 @@ const public_key_to_wif = require('../private/public_key_to_wif')
  * Generate a new cryptographically random EOS key pair.
  * @kind function
  * @name new_eos_keys
+ * @param {Uint8Array} [seed] A 32 byte array to seed a private key (seed < curve order n).
  * @returns {KeyPair} Key pair.
  * @example <caption>Ways to `import`.</caption>
  * ```js
@@ -37,8 +38,10 @@ const public_key_to_wif = require('../private/public_key_to_wif')
  * ```
  * The logged output will be an object containing EOS wif public & private keys.
  */
-async function new_eos_keys() {
-  const { private_key, public_key } = await generate_key_pair()
+async function new_eos_keys(seed) {
+  const private_key = seed ? seed : await random_bytes()
+  const public_key = await get_public_key(private_key)
+
   return {
     public_key: await public_key_to_wif(public_key),
     private_key: await private_key_to_wif(private_key)