+++ changelog.md
@@ -1,5 +1,33 @@
 # eos-ecc changelog
 
-## 1.0.0
+## Next
+
+### Patch
+
+- Updated dependency `universal-ecdsa`.
+
+- Added extra tests to private key to public key fn.
+
+### Major
+
+- Added `universal-ecdsa` package.
+
+- Removed signature validation.
+
+- Fix for invalid signature.
+
+### Minor
+
+- `new eos keys` function now contains a seed argument.
+
+- Removed `hard-coverage`.
+
+- Removed `isomorphic-secp256k1`.
+
+### Patch
+
+Updated dependencies.
+
+## 0.0.0-rc
 
 Initial Release.